weather-api/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==1.0.0
numpy==1.24.3
pandas3==0.0.1
Flask-Cors==4.0.0